+++ src/lib/data/queries.ts
@@ -1,5 +1,23 @@
 import { graphql } from "lib/gql";
 
+export const getCodeListQueryDocument = graphql(`
+  query getCodeListQuery {
+    codes(limit: 500, offset: 0, order_by: { id: desc }) {
+      id
+      contracts_aggregate {
+        aggregate {
+          count
+        }
+      }
+      account {
+        uploader: address
+      }
+      access_config_permission
+      access_config_addresses
+    }
+  }
+`);
+
 export const getCodeListByUserQueryDocument = graphql(`
   query getCodeListByUserQuery($walletAddr: String!) {
     codes(
@@ -9,7 +27,11 @@ export const getCodeListByUserQueryDocument = graphql(`
       order_by: { id: desc }
     ) {
       id
-      instantiated: contract_instantiated
+      contracts_aggregate {
+        aggregate {
+          count
+        }
+      }
       account {
         uploader: address
       }
@@ -23,7 +45,11 @@ export const getCodeListByIDsQueryDocument = graphql(`
   query getCodeListByIDsQuery($ids: [Int!]!) {
     codes(where: { id: { _in: $ids } }) {
       id
-      instantiated: contract_instantiated
+      contracts_aggregate {
+        aggregate {
+          count
+        }
+      }
       account {
         uploader: address
       }
@@ -121,6 +147,83 @@ export const getExecuteTxsCountByContractAddress = graphql(`
   }
 `);
 
+export const getMigrationHistoriesByContractAddress = graphql(`
+  query getMigrationHistoriesByContractAddress(
+    $contractAddress: String!
+    $offset: Int!
+    $pageSize: Int!
+  ) {
+    contract_histories(
+      where: { contract: { address: { _eq: $contractAddress } } }
+      order_by: { block: { timestamp: desc } }
+      limit: $pageSize
+      offset: $offset
+    ) {
+      code_id
+      account {
+        address
+      }
+      block {
+        height
+        timestamp
+      }
+      remark
+    }
+  }
+`);
+
+export const getMigrationHistoriesCountByContractAddress = graphql(`
+  query getMigrationHistoriesCountByContractAddress($contractAddress: String!) {
+    contract_histories_aggregate(
+      where: { contract: { address: { _eq: $contractAddress } } }
+    ) {
+      aggregate {
+        count
+      }
+    }
+  }
+`);
+
+export const getRelatedProposalsByContractAddress = graphql(`
+  query getRelatedProposalsByContractAddress(
+    $contractAddress: String!
+    $offset: Int!
+    $pageSize: Int!
+  ) {
+    contract_proposals(
+      where: { contract: { address: { _eq: $contractAddress } } }
+      order_by: { proposal_id: desc }
+      offset: $offset
+      limit: $pageSize
+    ) {
+      proposal {
+        title
+        status
+        voting_end_time
+        deposit_end_time
+        type
+        account {
+          address
+        }
+      }
+      proposal_id
+      resolved_height
+    }
+  }
+`);
+
+export const getRelatedProposalsCountByContractAddress = graphql(`
+  query getRelatedProposalsCountByContractAddress($contractAddress: String!) {
+    contract_proposals_aggregate(
+      where: { contract: { address: { _eq: $contractAddress } } }
+    ) {
+      aggregate {
+        count
+      }
+    }
+  }
+`);
+
 export const getContractListByCodeId = graphql(`
   query getContractListByCodeId($codeId: Int!, $offset: Int!, $pageSize: Int!) {
     contracts(
@@ -176,6 +279,54 @@ export const getCodeInfoByCodeId = graphql(`
       }
       access_config_permission
       access_config_addresses
+    }
+  }
+`);
+
+export const getTxsByContractAddress = graphql(`
+  query getTxsByContractAddress(
+    $contractAddress: String!
+    $offset: Int!
+    $pageSize: Int!
+  ) {
+    contract_transactions(
+      where: { contract: { address: { _eq: $contractAddress } } }
+      order_by: { transaction: { block: { timestamp: desc } } }
+      offset: $offset
+      limit: $pageSize
+    ) {
+      transaction {
+        hash
+        success
+        messages
+        account {
+          address
+        }
+        block {
+          height
+          timestamp
+        }
+        is_execute
+        is_ibc
+        is_instantiate
+        is_send
+        is_store_code
+        is_migrate
+        is_update_admin
+        is_clear_admin
+      }
+    }
+  }
+`);
+
+export const getTxsCountByContractAddress = graphql(`
+  query getTxsCountByContractAddress($contractAddress: String!) {
+    contract_transactions_aggregate(
+      where: { contract: { address: { _eq: $contractAddress } } }
+    ) {
+      aggregate {
+        count
+      }
     }
   }
 `);