+++ src/lib/pages/codes/components/CodeDescriptionCell.tsx
@@ -4,21 +4,18 @@ import { MdCheckCircle } from "react-icons/md";
 import { EditableCell } from "lib/components/table";
 import { MAX_CODE_DESCRIPTION_LENGTH } from "lib/data";
 import { useCodeStore } from "lib/hooks";
+import type { CodeLocalInfo } from "lib/stores/code";
 
 interface CodeDescriptionCellProps {
-  codeId: number;
-  description?: string;
+  code: CodeLocalInfo;
 }
 
-export const CodeDescriptionCell = ({
-  codeId,
-  description,
-}: CodeDescriptionCellProps) => {
+export const CodeDescriptionCell = ({ code }: CodeDescriptionCellProps) => {
   const toast = useToast();
   const { updateCodeInfo } = useCodeStore();
 
   const onSave = (inputValue?: string) => {
-    updateCodeInfo(codeId, { description: inputValue });
+    updateCodeInfo(code.id, code.uploader, inputValue);
     toast({
       title: "New Code Description saved",
       status: "success",
@@ -38,7 +35,7 @@ export const CodeDescriptionCell = ({
   };
   return (
     <EditableCell
-      initialValue={description}
+      initialValue={code.description}
       defaultValue="No Description"
       maxLength={MAX_CODE_DESCRIPTION_LENGTH}
       onSave={onSave}